\name{SurvivalClusteringTree-package}
\alias{SurvivalClusteringTree-package}
\alias{SurvivalClusteringTree}
\docType{package}
\title{\packageTitle{SurvivalClusteringTree}}
\description{\packageDescription{SurvivalClusteringTree}}
\section{Package Content}{\packageIndices{SurvivalClusteringTree}}
\author{\packageAuthor{SurvivalClusteringTree}}
\section{Maintainer}{\packageMaintainer{SurvivalClusteringTree}}
\keyword{package}
