% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival_tree.R
\name{plot_survival_tree}
\alias{plot_survival_tree}
\title{Visualize the Fitted Survival Tree}
\usage{
plot_survival_tree(survival_tree, cex = 0.75)
}
\arguments{
\item{survival_tree}{a fitted survival tree object.}

\item{cex}{numeric character expansion factor.}
}
\description{
Visualize the Fitted Survival Tree
}
